+++ inlang/source-code/ide-extension/src/utilities/settings/recommendation.ts
@@ -2,8 +2,9 @@ import { getSetting, updateSetting } from "../settings/index.js"
 import * as vscode from "vscode"
 import * as path from "node:path"
 import * as fs from "node:fs"
-import { getGitOrigin, telemetry } from "../../services/telemetry/implementation.js"
+import { telemetry } from "../../services/telemetry/implementation.js"
 import { parse, stringify, type CommentJSONValue } from "comment-json"
+import { getGitOrigin } from "./getGitOrigin.js"
 
 /**
  * Defines the structure of the extensions.json file.