+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/src/runtime/utils/external.ts
@@ -1,0 +1,13 @@
+export function isExternal(url: URL, currentUrl: URL, base: string) {
+	const absoluteBase = new URL(base, currentUrl).pathname
+
+	if (url.origin !== currentUrl.origin) {
+		return true
+	}
+
+	if (!url.pathname.startsWith(absoluteBase)) {
+		return true
+	}
+
+	return false
+}