+++ inlang/source-code/sdk/src/loadProject.ts
@@ -43,39 +43,16 @@ import { createNodeishFsWithWatcher } from "./createNodeishFsWithWatcher.js"
 import { maybeCreateFirstProjectId } from "./migrations/maybeCreateFirstProjectId.js"
 
 import { capture } from "./telemetry/capture.js"
+import { identifyProject } from "./telemetry/groupIdentify.js"
 
 const settingsCompiler = TypeCompiler.Compile(ProjectSettings)
-
-/**
- * Creates an inlang instance.
- *
- * @param projectPath - Absolute path to the inlang settings file.
- * @param @deprecated nodeishFs - Filesystem that implements the NodeishFilesystemSubset interface.
- * @param _import - Use `_import` to pass a custom import function for testing,
- *   and supporting legacy resolvedModules such as CJS.
- *
- */
-export async function loadProject(args: {
-	projectPath: string
-	nodeishFs: Repository["nodeishFs"]
-	/**
-	 * The app id is used to identify the app that is using the SDK.
-	 *
-	 * We use the app id to group events in telemetry to answer questions
-	 * like "Which apps causes these errors?" or "Which apps are used more than others?".
-	 *
-	 * @example
-	 * 	appId: "app.inlang.badge"
-	 */
-	appId?: string
-	_import?: ImportFunction
-}): Promise<InlangProject>
 
 /**
  * @param projectPath - Absolute path to the inlang settings file.
  * @param repo - An instance of a lix repo as returned by `openRepository`.
  * @param _import - Use `_import` to pass a custom import function for testing,
  *   and supporting legacy resolvedModules such as CJS.
+ * @param appId - The app id to use for telemetry e.g "app.inlang.badge"
  *
  */
 export async function loadProject(args: {
@@ -83,14 +60,6 @@ export async function loadProject(args: {
 	repo: Repository
 	appId?: string
 	_import?: ImportFunction
-}): Promise<InlangProject>
-
-export async function loadProject(args: {
-	projectPath: string
-	repo?: Repository
-	appId?: string
-	_import?: ImportFunction
-	nodeishFs?: Repository["nodeishFs"]
 }): Promise<InlangProject> {
 	const projectPath = normalizePath(args.projectPath)
 
@@ -111,15 +80,7 @@ export async function loadProject(args: {
 		)
 	}
 
-	let fs: Repository["nodeishFs"]
-	if (args.nodeishFs) {
-		// TODO: deprecate
-		fs = args.nodeishFs
-	} else if (args.repo) {
-		fs = args.repo.nodeishFs
-	} else {
-		throw new LoadProjectInvalidArgument(`Repo missing from arguments.`, { argument: "repo" })
-	}
+	const fs = args.repo.nodeishFs
 
 	const nodeishFs = createNodeishFsWithAbsolutePaths({
 		projectPath,
@@ -545,6 +506,13 @@ export async function loadProject(args: {
 						(plugin) => plugin.loadMessages !== undefined
 					)
 
+					const monkePatchedFileRead = fsWithWatcher.readFile
+
+					fsWithWatcher.readFile = async function (filename: string) {
+						console.log(filename)
+						return monkePatchedFileRead(filename)
+					} as any
+
 					// TODO #1844 FINK check error handling for plugin load methods (triggered by file change) -> move to separate ticket
 					loadMessages(fsWithWatcher, messagesQuery, settings()!, loadMessagePlugin)
 				}
@@ -625,6 +593,14 @@ export async function loadProject(args: {
 		if (projectId && projectLoadedCapturedAlready === false) {
 			projectLoadedCapturedAlready = true
 			// TODO ensure that capture is "awaited" without blocking the the app from starting
+			await identifyProject({
+				projectId,
+				properties: {
+					// using the id for now as a name but can be changed in the future
+					// we need at least one property to make a project visible in the dashboard
+					name: projectId,
+				},
+			})
 			await capture("SDK loaded project", {
 				projectId,
 				properties: {
@@ -638,6 +614,7 @@ export async function loadProject(args: {
 		}
 
 		return {
+			id: projectId,
 			installed: {
 				plugins: createSubscribable(() => installedPlugins()),
 				messageLintRules: createSubscribable(() => installedMessageLintRules()),