+++ inlang/source-code/website/src/interface/marketplace/categoryHeaders/categoryHeros/appHeader.tsx
@@ -5,7 +5,7 @@ import { For, Match, Show, Switch } from "solid-js"
 
 const AppHeader = () => {
 	const getProducts = () => [
-		"app.inlang.editor",
+		"app.inlang.finkLocalizationEditor",
 		"app.inlang.ideExtension",
 		"app.parrot.figmaPlugin",
 	]
@@ -13,7 +13,7 @@ const AppHeader = () => {
 	// const data: Array<DataItem> = [
 	// 	{
 	// 		title: "Fink – Message editor",
-	// 		link: "/m/tdozzpar/app-inlang-editor",
+	// 		link: "/m/tdozzpar/app-inlang-finkLocalizationEditor",
 	// 		description:
 	// 			"The Fink message editor enables you to work with globalized codebases in your browser...",
 	// 		pricing: "Free Beta",
@@ -71,7 +71,7 @@ const AppHeader = () => {
 							>
 								<div class="w-full h-[200px] overflow-hidden">
 									<Switch>
-										<Match when={manifest.id === "app.inlang.editor"}>
+										<Match when={manifest.id === "app.inlang.finkLocalizationEditor"}>
 											<img
 												class="group-hover:scale-105 h-full w-full bg-surface-400 object-cover transition-all duration-500"
 												src={"/images/fink-cover.png"}
@@ -141,9 +141,6 @@ const AppHeader = () => {
 					}}
 				</For>
 			</div>
-			<h2 class="pb-4 border-t-surface-200 text-xl font-medium tracking-tight text-surface-900">
-				All Apps
-			</h2>
 		</>
 	)
 }